# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Docker API",
    "description": "A production-ready Go REST API that exposes a simplified wrapper around the Docker Engine API. Supports container lifecycle management and resource metrics.",
    "version": "1.1.0",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/example/docker-api"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/system/info": {
      "get": {
        "summary": "System info",
        "description": "Returns Docker daemon and host info (version, OS, driver, memory, CPUs, etc.).",
        "operationId": "getSystemInfo",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "Docker and host information (types.Info)",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "500": {
            "description": "Failed to get system info",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/system/disk": {
      "get": {
        "summary": "Disk usage",
        "description": "Returns disk usage for images, containers, volumes, and build cache.",
        "operationId": "getSystemDisk",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "Disk usage (types.DiskUsageReport)",
            "content": { "application/json": { "schema": { "type": "object" } } }
          },
          "500": {
            "description": "Failed to get disk usage",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
    "/volumes": {
      "get": {
        "summary": "List volumes",
        "description": "Returns all Docker volumes.",
        "operationId": "listVolumes",
        "tags": ["Volumes"],
        "responses": {
          "200": {
            "description": "List of volumes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VolumeSummary" }
                }
              }
            }
          },
          "500": { "description": "Failed to list volumes", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "summary": "Create volume",
        "description": "Creates a new volume. Body: name (required), driver, labels, driver_opts.",
        "operationId": "createVolume",
        "tags": ["Volumes"],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateVolumeRequest" } } }
        },
        "responses": {
          "201": { "description": "Volume created", "content": { "application/json": { "schema": { "type": "object" } } } },
          "400": { "description": "Invalid body or missing name", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Failed to create volume", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/volumes/prune": {
      "post": {
        "summary": "Prune volumes",
        "description": "Removes unused volumes.",
        "operationId": "pruneVolumes",
        "tags": ["Volumes"],
        "responses": {
          "200": { "description": "Prune report", "content": { "application/json": { "schema": { "type": "object" } } } },
          "500": { "description": "Failed to prune", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/volumes/{name}": {
      "get": {
        "summary": "Inspect volume",
        "operationId": "inspectVolume",
        "tags": ["Volumes"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Volume details", "content": { "application/json": { "schema": { "type": "object" } } } },
          "400": { "description": "Missing volume name", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Failed to inspect", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "summary": "Remove volume",
        "operationId": "removeVolume",
        "tags": ["Volumes"],
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "force", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Volume removed" },
          "500": { "description": "Failed to remove", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/networks": {
      "get": {
        "summary": "List networks",
        "operationId": "listNetworks",
        "tags": ["Networks"],
        "responses": {
          "200": { "description": "List of networks", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/NetworkSummary" } } } } },
          "500": { "description": "Failed to list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "summary": "Create network",
        "operationId": "createNetwork",
        "tags": ["Networks"],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateNetworkRequest" } } } },
        "responses": {
          "201": { "description": "Network created" },
          "400": { "description": "Invalid body", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Failed to create", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/networks/prune": {
      "post": {
        "summary": "Prune networks",
        "operationId": "pruneNetworks",
        "tags": ["Networks"],
        "responses": {
          "200": { "description": "Prune report" },
          "500": { "description": "Failed to prune", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/networks/{id}/connect": {
      "post": {
        "summary": "Connect container to network",
        "operationId": "connectNetwork",
        "tags": ["Networks"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "required": ["container"], "properties": { "container": { "type": "string" } } } } } },
        "responses": {
          "204": { "description": "Connected" },
          "400": { "description": "Invalid body", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/networks/{id}/disconnect": {
      "post": {
        "summary": "Disconnect container from network",
        "operationId": "disconnectNetwork",
        "tags": ["Networks"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "required": ["container"], "properties": { "container": { "type": "string" }, "force": { "type": "boolean" } } } } } },
        "responses": {
          "204": { "description": "Disconnected" },
          "400": { "description": "Invalid body", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/networks/{id}": {
      "get": {
        "summary": "Inspect network",
        "operationId": "inspectNetwork",
        "tags": ["Networks"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Network details", "content": { "application/json": { "schema": { "type": "object" } } } },
          "400": { "description": "Missing network ID", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "summary": "Remove network",
        "operationId": "removeNetwork",
        "tags": ["Networks"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "204": { "description": "Network removed" },
          "400": { "description": "Missing network ID", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VolumeSummary": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "driver": { "type": "string" },
          "mountpoint": { "type": "string" },
          "labels": { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
      "CreateVolumeRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "driver": { "type": "string" },
          "labels": { "type": "object", "additionalProperties": { "type": "string" } },
          "driver_opts": { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
      "NetworkSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "driver": { "type": "string" },
          "scope": { "type": "string" },
          "containers": { "type": "object", "additionalProperties": true }
        }
      },
      "CreateNetworkRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "driver": { "type": "string" },
          "options": { "type": "object", "additionalProperties": { "type": "string" } },
          "ipam": { "type": "object" }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "failed to start container"
          }
        }
      }
    }
  }
}